refactor: consolidate tape and store modules
