+++ src/bub/builtin/model_runner.py
@@ -28,7 +28,6 @@ from pydantic import TypeAdapter, ValidationError
 
 from bub.builtin.codex_provider import OpenaiCodexProvider, should_use_openai_codex_provider
 from bub.builtin.settings import AgentSettings, ModelCandidate
-from bub.builtin.tape import Tape
 from bub.errors import BubError, ErrorKind
 from bub.hooks.interception import (
     AgentHooks,
@@ -37,6 +36,7 @@ from bub.hooks.interception import (
     LlmCallResult,
 )
 from bub.streaming import AsyncStreamEvents, StreamEvent, StreamState
+from bub.tape import Tape
 from bub.tools import Tool, ToolContext, ToolExecutor
 
 CONTEXT_LENGTH_PATTERNS = re.compile(
